Move helper function from test to shared test trait
