+++ tests/phpunit/CRMTraits/Import/ParserTrait.php
@@ -9,6 +9,8 @@
  +--------------------------------------------------------------------+
  */
 
+use Civi\Api4\DedupeRuleGroup;
+
 /**
  * Trait ParserTrait
  *
@@ -133,6 +135,40 @@ trait CRMTraits_Import_ParserTrait {
     $this->userJobID = $form->getUserJobID();
     // This gets reset in DataSource so re-do....
     $_SESSION['_' . $form->controller->_name . '_container']['values'] = $values;
+  }
+
+  /**
+   * Enhance field such that any combo of the custom field & first/last name is enough.
+   *
+   * @noinspection PhpUnhandledExceptionInspection
+   */
+  protected function addToDedupeRule(): void {
+    $this->createCustomGroupWithFieldOfType(['extends' => 'Contact']);
+    $dedupeRuleGroup = DedupeRuleGroup::get()
+      ->addWhere('name', '=', 'IndividualUnsupervised')
+      ->addSelect('id', 'threshold')
+      ->execute()
+      ->first();
+    $this->assertEquals(10, $dedupeRuleGroup['threshold']);
+    $dedupeRuleGroupID = $this->ids['DedupeRule']['unsupervised'] = $dedupeRuleGroup['id'];
+    $this->callAPISuccess('Rule', 'create', [
+      'dedupe_rule_group_id' => $dedupeRuleGroupID,
+      'rule_weight' => 5,
+      'rule_table' => $this->getCustomGroupTable(),
+      'rule_field' => $this->getCustomFieldColumnName('text'),
+    ]);
+    $this->callAPISuccess('Rule', 'create', [
+      'dedupe_rule_group_id' => $dedupeRuleGroupID,
+      'rule_weight' => 5,
+      'rule_table' => 'civicrm_contact',
+      'rule_field' => 'first_name',
+    ]);
+    $this->callAPISuccess('Rule', 'create', [
+      'dedupe_rule_group_id' => $dedupeRuleGroupID,
+      'rule_weight' => 5,
+      'rule_table' => 'civicrm_contact',
+      'rule_field' => 'last_name',
+    ]);
   }
 
 }
